+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Plan [1.18.3]
+
+22 Nov, 2021
+
+### Changes
+- Provide plans for signup and upgrade flow separately
+
 ## Network [1.16.0]
 
 22 Nov, 2021